+++ extensions/ql-vscode/src/data-extensions-editor/data-extensions-editor-module.ts
@@ -21,7 +21,7 @@ import { extLogger } from "../common/logging/vscode";
 import { isQueryLanguage } from "../common/query-language";
 import { setUpPack } from "./external-api-usage-queries";
 import { DisposableObject } from "../common/disposable-object";
-import { ModelDetailsPanel } from "./model-details/model-details-panel";
+import { ModelDetailsPanel } from "./model-details/methods-usage-panel";
 import { Mode } from "./shared/mode";
 import { showResolvableLocation } from "../databases/local-databases/locations";
 import { Usage } from "./external-api-usage";
