+++ extensions/ql-vscode/src/data-extensions-editor/model-details/methods-usage-data-provider.ts
